Fixed non empty dir removal when sqlite project testing

+++ sqlite/src/test/java/ua/pp/msk/yum/sqlite/primary/jdbc/InitDbTest.java
@@ -67,7 +67,9 @@ public class InitDbTest {
                 Files.delete(tdb);
             }
             if (Files.exists(parentDb)) {
-                Files.delete(parentDb);
+                if (parentDb.toFile().list().length == 0) {
+                    Files.delete(parentDb);
+                }
             }
           
         } catch (IOException ex) {
